backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.6.3
motor==3.3.2
python-multipart==0.0.9
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
dlib==19.24.2
reportlab==4.1.0
pydantic-settings==2.3.4
pillow==10.3.0
python-dotenv==1.0.1
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models
